Separate bgtask
