+++ src/ai/backend/common/BUILD
@@ -26,6 +26,7 @@ python_distribution(
     dependencies=[
         ":src",
         "src/ai/backend/common/auth:src",
+        "src/ai/backend/common/bgtask:src",
         "src/ai/backend/common/contexts:src",
         "src/ai/backend/common/data/session:src",
         "src/ai/backend/common/data/vfolder:src",
