+++ src/ai/backend/common/bgtask/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
