+++ src/ai/backend/manager/api/context.py
@@ -13,7 +13,7 @@ from ai.backend.manager.service.base import ServicesContext
 from ai.backend.manager.services.processors import Processors
 
 if TYPE_CHECKING:
-    from ai.backend.common.bgtask import BackgroundTaskManager
+    from ai.backend.common.bgtask.bgtask import BackgroundTaskManager
     from ai.backend.common.events.events import EventDispatcher, EventProducer
     from ai.backend.common.plugin.hook import HookPluginContext
     from ai.backend.common.plugin.monitor import ErrorPluginContext, StatsPluginContext
