+++ src/ai/backend/manager/container_registry/base.py
@@ -24,7 +24,7 @@ import sqlalchemy as sa
 import trafaret as t
 import yarl
 
-from ai.backend.common.bgtask import ProgressReporter
+from ai.backend.common.bgtask.bgtask import ProgressReporter
 from ai.backend.common.docker import (
     ImageRef,
     arch_name_aliases,
