+++ src/ai/backend/manager/models/gql_models/agent.py
@@ -20,7 +20,7 @@ from sqlalchemy.engine.row import Row
 from sqlalchemy.ext.asyncio import AsyncConnection as SAConnection
 
 from ai.backend.common import msgpack, redis_helper
-from ai.backend.common.bgtask import ProgressReporter
+from ai.backend.common.bgtask.bgtask import ProgressReporter
 from ai.backend.common.json import dump_json_str, load_json
 from ai.backend.common.types import (
     AccessKey,
