+++ src/ai/backend/manager/services/model_serving/services/model_serving.py
@@ -10,7 +10,7 @@ from sqlalchemy.orm import selectinload
 from sqlalchemy.orm.exc import NoResultFound
 from yarl import URL
 
-from ai.backend.common.bgtask import BackgroundTaskManager, ProgressReporter
+from ai.backend.common.bgtask.bgtask import BackgroundTaskManager, ProgressReporter
 from ai.backend.common.events.events import (
     EventDispatcher,
     EventHandler,
