+++ src/ai/backend/manager/services/processors.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import Self
 
-from ai.backend.common.bgtask import BackgroundTaskManager
+from ai.backend.common.bgtask.bgtask import BackgroundTaskManager
 from ai.backend.common.etcd import AsyncEtcd
 from ai.backend.common.events.events import EventDispatcher
 from ai.backend.common.plugin.monitor import ErrorPluginContext
